+++ zenml/steps/trainer/tensorflow_trainers/__init__.py
@@ -11,5 +11,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
-
-from .tf_base_trainer import TFBaseTrainerStep
